requirements.txt
requests
python-dotenv
python-crontab
msal
pysnmp
python-gvm
pyzabbix
python-nmap
pymysql